* unstable:
  qfc: increase precision

+++ qfunctionconfigurator/functionconfig.h
@@ -21,7 +21,7 @@ class Map {
 private:
     static constexpr int value_count = 10000;
     using num = double;
-    using integral = std::uint16_t;
+    using integral = std::uint32_t;
     static constexpr integral integral_max = std::numeric_limits<integral>::max();
 
     struct State {
